requirements.txt

flask==2.3.3
flask-cors==4.0.0

# TensorFlow (CPU version untuk Railway) - simplified
tensorflow-cpu==2.13.0
# If above fails, try: tensorflow==2.13.0


Pillow==10.0.0
numpy==1.24.3


gunicorn==21.2.0


gevent==23.9.1


werkzeug==2.3.7